requirements.txt
streamlit
Pillow
requests
beautifulsoup4
lxml
python-dotenv
diffusers
transformers
torch
torchvision
accelerate